pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "multi-agent-consensus"
version = "0.1.0"
description = "Framework for 3-agent consensus validation with hard gates for Claude Code"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Nick Krzemienski", email = "nick@krzemienski.com"},
]
keywords = ["claude", "ai", "consensus", "validation", "multi-agent", "quality-assurance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
consensus = "consensus.cli:cli"

[project.urls]
Homepage = "https://github.com/krzemienski/multi-agent-consensus"
Repository = "https://github.com/krzemienski/multi-agent-consensus"
Issues = "https://github.com/krzemienski/multi-agent-consensus/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
